Fix: edit or remove default notes list

+++ app/Http/Controllers/NoteListController.php
@@ -12,14 +12,17 @@ class NoteListController extends Controller
     public function index(Request $request)
     {
         $user = $request->user();
-        $listsOfNotes = $user->lists;
-        // $listsOfNotes = NoteList::where('user_id', $user->id)->with('notes')->get();
+        $listsOfNotes = NoteList::where('user_id', $user->id)->with('notes')->get();
         return response()->json($listsOfNotes);
     }
 
     public function store(NoteListFormRequest $request)
     {
         $user = $request->user();
+
+        if ($request->only('title') == "All Notes (default)") {
+            abort(403, "Invalid name for a list of notes");
+        }
 
         $newNoteList = array_merge($request->all(), ['user_id' => $user->id]);
         $noteList = NoteList::create($newNoteList);
@@ -33,7 +36,7 @@ class NoteListController extends Controller
 
         $this->authorize('isAuthorized', [$noteList, $user]);
 
-        return response()->json($noteList);
+        return response()->json($noteList->with('notes')->where('id', $noteList->id)->get()[0]);
     }
 
     public function update(NoteListFormRequest $request,  $id)
@@ -42,6 +45,10 @@ class NoteListController extends Controller
         $noteList = NoteList::find($id);
 
         $this->authorize('isAuthorized', [$noteList, $user]);
+
+        if ($noteList->title == "All Notes (default)" && $noteList->description ==  "Note list created by default") {
+            abort(403, "It is not possible to update the list of notes by default");
+        }
 
         $noteList->update($request->all());
         return response()->json($noteList);
@@ -53,6 +60,10 @@ class NoteListController extends Controller
         $noteList = NoteList::find($id);
 
         $this->authorize('isAuthorized', [$noteList, $user]);
+
+        if ($noteList->title == "All Notes (default)" && $noteList->description ==  "Note list created by default") {
+            abort(403, "It is not possible to delete the list of notes by default");
+        }
 
         $noteList->delete();
         return response()->json($noteList);
